projectorscreen/pyproject.toml
[tool.poetry]
name = "projectorscreen"
version = "0.1.0"
description = ""
authors = ["Mariusz Kabała <mariusz@kabala.waw.pl>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.3"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
